Add LICENSE support to plugins
This will allow you to link to a text document that will display a LICENSE for the plugin before you are allowed to install it.

For plugin authors:
It's only necessary to supply the key "license", with the value being the location to the license itself on the web.

Example:
```
  "foo": {
        "MANIFEST": "foo.json",
        "name": "Foobar Baz",
        "icon": "icons/foo.png",
        "description": "The best plugin in the world.",
        "license": "https://raw.githubusercontent.com/foo/bar/LICENSE"
        }
```