Add cffi to pin list, pin others
CFFI is a dep of `pygit2`, but we are pinning it here to help avoid the annoyance of having to remove it and then install it again with `pygit2`.

Pinning the others to avoid this cat and mouse game. iocage should be in a virtualenv if using GitHub anyways. As always, you're recommended to use the `Makefile` and not `setup.py` directly.