More helpful message when snapshot already exists
When trying to iocage -f list to transition from an old non-python iocage installation, if the snapshot that is being created as part of the upgrade process already exists, a generic error will be printed and there is no way (that I know of) to get out of that situation by using iocage alone. By providing the exact snapshot that already exists, the user can e.g. choose to zfs rename -r said snapshot and thus overcome the impasse.