Do not falsely report zfs exceptions when creating template
Currently, when converting a jail into a template, if there are any errors
setting zfs options or renaming datasets, iocage will report that a template
of the same name exists.

This changes the behavior to throw an exception when we encounter this case,
like the opposite case (template -> jail) does.