Correctly create aliases for shared IP jails that don't specify iface (#28)
* Correctly create aliases for shared IP jails that don't specify iface

It seems jail(8) behavior may have changed, and this is now required for jail to create the aliases. We do a best effort attempt at using the default interface if the user doesn't supply one.

The idea is that if they didn't supply one, they may have already set the alias up themselves, so iocage will do nothing in this case.

Ticket: #46245

* Remove duplicate variable

This would be fatal possibly.

* Let's use the right variable

Possibly shadowing before anyways, this is better.