Bug fix for always creating jails with name and count flags
This commit fixes a bug where we were not able to create count jails when a valid uuid was supplied as a name. Apart from that this commit enforces that count flag cannot be used with uuid flag and removes force option from create command as it is not used anywhere.
Ticket: #61551