Bug fix for template jails
This commit fixes a bug which prevented creating a template jail at the same time the jail was being created for the first time. The dataset was readonly at that point and system tried to complete the relevant configuration files but failed to do so because of the readonly property set to true.
Ticket: #62190