Check Release exists
This commit makes sure we check the release exists before destroying it. If it doesn't, we raise an appropriate exception outlining that it does not exist. It also introduces a new class Resource which we can use to manipulate resources in iocage. However much thought is still needed on how we should proceed with it in the future.
Ticket: #63810