vnet_default_interface per bridge
Add ability to specify a third item in the
interfaces 'tuple', e.g. 'vnet0:bridge0:bge1'.

vnet_default_interface is still respected when
the third item is not present. An exception is
logged if the interface does not exist at parameter
set time or jail start time, just as
vnet_default_interfaces is checked.