Deprecate the mac_prefix property
With the new MAC-generation algorithm, the static mac_prefix property no
longer applies. Continuing to use the prefix, taken from the first 3
bytes of the host default interface, would result in MAC collision,
because the new MAC-generation algorithm uses the last 3 bytes of
the host default interface as a suffix.