Check if jail is running before destroying it and if, don't attempt t… (#831)
* destroy a jail only if it is not running

* destroy a jail only if it is not running

* call destroy_jail to destroy_release

This will call destroy_jail per child of the release and
with it the check if the child is still running, so that the
child only gets destroyed if it is not running. While this
introduces some effect on performance, it will only occurr
when a release has got dependent jails. When destroy_release
is called with --recursive and -f the children will be
stopped and destroyed afterwards, so that the release can be
destroyed anyways.

* fix indentation

* remove unnecessary return

* add callback