Warn users for bad perms
This commit adds a warning for users who have accidentally/deliberately changed ownership/permissions for /usr/lib inside jail. The warning appears when the user tries iocage console, we don't error out but just give the user a friendly notice that there is a good chance login(1) is not going to like the changed perms and so he/she should have that corrected.
Ticket: #NAS-101560