Correctly set base jail release
This commit fixes an issue where we didn't update the release property of a base jail when we updated it if the property in config did not match the supplied one when updating. How this is an issue is
that the first time we get a release which has not been updated will have the release as 11.2-RELEASE. First time we create a jail and update it to say 11.2-RELEASE-p4, it would be reflected as desired
because the release passed on to update was 11.2-RELEASE which matched with the jail's prop. Moving on the release specified to update would always be 11.2-RELEASE but the prop's value has changed so it would
not match and hence the property would not be updated.

Ticket: #NAS-101815