Renamed plugins won't upgrade/update
This commit fixes a bug where renamed plugins won't upgrade/update because they'll beleive that their plugin name is actually the jail's id which is obviously not true if the jail has been renamed. We have a prop to keep track of plugin name and that will take care of this now.