Improve capability to guess plugin's name
Before we had introduced a property for determining plugin name of a plugin jail, we had no robust way to guess which plugin jail referenced which plugin in the manifest. While making a shift to property, this commit adds ability to search for plugin's json file which is supposed to be present in the jail and retrieve name property from it.