Correctly list admin portal for DHCP/Static IP plugin jails
This commit introduces changes to ensure that for DHCP/Static ip plugin jails, we don't list the host's ip address in the admin portal. For NAT based jails, it also ensures that the correct nat ports are specified for the plugin jail. However, if the nat_forwards haven't been set by the plugin dev or the end user, we only show the system ip then in that case.