Change devfs ruleset handling so that configured rulesets != 4 are cloned/copied into dynamic ones.  This makes devfs rule handling more symmetrical to what is done when the default ruleset 4 is configured (which in fact never applies devfs ruleset 4, but creates an iocage specific dynamic ruleset instead - which can be quite confusing).
As a result, this addresses the problem of non-dynamic rulesets being
removed on `iocage stop` raised in
https://github.com/iocage/iocage/issues/952.

This also makes iocage fail to start a jail if the configured devfs ruleset
is not available - beforehand it would start with a default ruleset in this
case, which can have severe unwanted side effects.

Finally, this sets the minimum dynamically assigned devfs rule id to 1000 to
reserve the lower ids for static configuration by the admin in devfs.rules
(this is mostly for convenience).