Only update release property on update of jails
This commit fixes an issue where iocage updated the complete configuration of jails on update for updating release property. This results in undesired behaviour with jails having partial configuration resulting in complete configuration and some properties are different when jail is running, i.e devfs_ruleset, that is changed as well.

Closes #1130