Make sure valid mac prefix is used when creating jails
This commit fixes an issue where if mac_prefix is not explicitly defined and the one set in defaults.json is an invalid one, we move forward using the invalid one. This is not desired, so if mac_prefix is not specified by the user and the one found in defaults.json is invalid, we on jail creation use a valid one then logging a warning. A more strict approach has not been taken in this case because it would require users to correct the mac prefix in defaults.json ( which would have wrong value if was created at a time before iocage patched the issue ) and any existing jails depending on that will inherit that change which in many environments will not be desired if the env depends on the mac address of the jail.