iocage_lib: fetch freebsd-update from releng/${version} instead
Fetching from the release/ tag means that we get the as-released
freebsd-update.sh, but with the caveat that we'll never get any
follow-up EN/SA that apply to freebsd-update.

Switch to the releng/ branch of the target release so we can pick up
any post-release advancements. As an added bonus, we don't need to
bother with the .0 concatenation since there is a 1:1 transformation
between X.Y-RELEASE -> releng/X.Y instead of the release/ legacy
naming convention.